+++ pkg/commands/ip.go
@@ -7,8 +7,8 @@ import (
 	"net/url"
 	"time"
 
-	"github.com/icarus-go/net-speed/pkg/command"
-	"github.com/icarus-go/net-speed/pkg/ipinfo"
+	"github.com/icarus-go/netspeed/pkg/command"
+	"github.com/icarus-go/netspeed/pkg/ipinfo"
 )
 
 // IPCommand IP 检测命令